requirement.txt
langchain
streamlit
google-generativeai
chromadb
python-dotenv
pyPDF2